add new example with updated Makefile
Added exception handlers that can be used in combination with gdb and the armv7m-vectstate.gdb macro to diagnose stack and heap overflow. Add optional feature to "color" memory with patterns to make it easier to take how much stack and heap are actually used.  Tested with the latest arm-eabi-gcc / gdb version 8.3.0